requirements.txt
fastapi
python-dotenv
line-bot-sdk
google-genai
uvicorn
pydantic-settings